backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
google-generativeai==0.3.2
pypdf2==3.0.1
python-docx==1.1.0
chromadb==0.4.22
sentence-transformers==2.2.2
requests==2.31.0
pydantic==2.5.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-dotenv==1.0.0
jinja2==3.1.2
aiofiles==23.2.1
langchain==0.1.0
langchain-community==0.0.20
faiss-cpu==1.7.4
numpy==1.24.3
huggingface-hub==0.19.4
transformers==4.35.2
feedparser==6.0.10
jinja2==3.1.2
email-validator==2.1.0
